backend_campus_connect/requirements.txt
asgiref
Django
djangorestframework
django-cors-headers
djoser
django-filter
psycopg2-binary
python-dotenv
PyJWT
pytz
sqlparse
djangorestframework-simplejwt
django-environ
whitenoise
django-webpack-loader
django-debug-toolbar
black
flake8
gunicorn
requests
pillow
dotenv
django-redis     
channels
channels_redis
redis
daphne
requests
decouple
cloudinary
dj_database_url
python-decouple